Added Exceptions and Validations

+++ src/main/java/com/crio/coderhack/services/UserService.java
@@ -5,6 +5,9 @@ import com.crio.coderhack.dtos.UserRequestDTO;
 import com.crio.coderhack.dtos.UserResponseDTO;
 import com.crio.coderhack.entities.Badge;
 import com.crio.coderhack.entities.User;
+import com.crio.coderhack.exceptions.InvalidScoreException;
+import com.crio.coderhack.exceptions.UserAlreadyExistsException;
+import com.crio.coderhack.exceptions.UserNotFoundException;
 import com.crio.coderhack.repositories.IUserRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +21,6 @@ import java.util.Optional;
 @Service
 public class UserService {
 
-    @Autowired
     private final IUserRepository userRepository;
     @Autowired
     private ModelMapper modelMapper;
@@ -27,17 +29,18 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    //TODO: Add Exceptions and checks
-
     public UserResponseDTO getUser(String userId){
         Long id = Long.parseLong(userId);
         Optional<User> user = userRepository.findById(id);
+        if (user.isEmpty()){
+            throw new UserNotFoundException("User Not Found!");
+        }
         return modelMapper.map(user, UserResponseDTO.class);
     }
 
     public List<UserResponseDTO> getUsers(){
         List<User> users = userRepository.findAll();
-        users.sort((a, b) -> Integer.compare(a.getScore(), b.getScore()));
+        users.sort((a, b) -> Integer.compare(b.getScore(), a.getScore()));
         List<UserResponseDTO> userResponseDTOS = new ArrayList<>();
         for(User user: users){
             UserResponseDTO userResponseDTO = modelMapper.map(user, UserResponseDTO.class);
@@ -47,6 +50,10 @@ public class UserService {
     }
 
     public UserResponseDTO registerUser(UserRequestDTO userRequestDTO){
+        Optional<User> existingUser = userRepository.findByUsername(userRequestDTO.getUsername());
+        if(existingUser.isPresent()){
+            throw new UserAlreadyExistsException("User Already Exists!");
+        }
         User user = new User(userRequestDTO.getUsername());
         long totalUsers = (long) userRepository.findAll().size();
         user.setId(totalUsers+1);
@@ -58,6 +65,14 @@ public class UserService {
         Long id = Long.parseLong(userId);
         Badge badge = getBadge(scoreRequestDTO);
         Optional<User> user = userRepository.findById(id);
+        if (user.isEmpty()) {
+            throw new UserNotFoundException("User not found");
+        }
+
+        int score = scoreRequestDTO.getScore();
+        if (score < 1 || score > 100) {
+            throw new InvalidScoreException("Score must be between 1 and 100");
+        }
         user.get().setScore(scoreRequestDTO.getScore());
         user.get().addBadge(badge);
         userRepository.save(user.get());
@@ -66,6 +81,10 @@ public class UserService {
 
     public void deleteUser(String userId){
         Long id = Long.parseLong(userId);
+        Optional<User> user = userRepository.findById(id);
+        if (user.isEmpty()) {
+            throw new UserNotFoundException("User not found");
+        }
         userRepository.deleteById(id);
     }
